Add SQLite-backed session store
